Migrate homepage, session workflow to Actix
Some initial thoughts:

Actix has fewer (no?) macros, so it's not magical. It's pretty clear
what it's doing and why.

The Handler contract can be frustrating to satisfy and the error
messages are somewhat cryptic. I found it somewhat challenging to make
sure it was done correctly.

I'm currently halting between two ideological positions, the first is
using Actix to plumb my application state and configuration into the
handlers (using Data) and the second is using some globals to do it out
of band.

- Data is advantaged in that the lifting is already done for me. There's
  really no additional effort.
- Data is disadvantaged in that it makes the handler's function
  arguments rather long. I'm not a fan of that.
- Conversely, globals are advantaged in that it keeps handler function
  arguments very short.

Currently the template engine is the only thing stuffed into a global to
emulate the API of Rocket regarding templates, but it's not completely
similar. In so doing I've noticed a subtlety of Rocket that I hadn't
noticed before, in that Template return type for handlers gets fed into
another Responder which pulls the template context out of the HTTP
request context - so it's effectively using Data again, only with less
boilerplate and ceremony.

If I can reduce the boilerplate in the ArcSwapOption pattern that may be
the way to go, or the other thing I've been considering is packaging up
the application state into a single context struct, so a single argument
in the handler gets everything - templates, database connection pool,
the works.

Something to ponder.